BATM-2237 move bech32 util classes
move bitcoincash specific implementation to bitcoincash package,
move generic implementation from litecoin to cashutil package
