+++ server_extensions_api/src/main/java/com/generalbytes/batm/server/coinutil/Bech32.java
@@ -1,183 +1,169 @@
-/*************************************************************************************
- * Copyright (C) 2014-2020 GENERAL BYTES s.r.o. All rights reserved.
+/*
+ * Copyright 2018 Coinomi Ltd
  *
- * This software may be distributed and modified under the terms of the GNU
- * General Public License version 2 (GPL2) as published by the Free Software
- * Foundation and appearing in the file GPL2.TXT included in the packaging of
- * this file. Please note that GPL2 Section 2[b] requires that all works based
- * on this software must also be made publicly available under the terms of
- * the GPL2 ("Copyleft").
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- * Contact information
- * -------------------
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
- * GENERAL BYTES s.r.o.
- * Web      :  http://www.generalbytes.com
- *
- ************************************************************************************/
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.generalbytes.batm.server.coinutil;
 
-import java.util.HashMap;
-import java.util.Map;
+import java.util.Arrays;
+import java.util.Locale;
 
+/** taken from litecoinj */
 public class Bech32 {
+    /** The Bech32 character set for encoding. */
+    private static final String CHARSET = "qpzry9x8gf2tvdw0s3jn54khce6mua7l";
 
-    public static final String CHARSET = "qpzry9x8gf2tvdw0s3jn54khce6mua7l";
-    private static final Long[] POLYMOD_GENERATORS = new Long[] {
-        Long.parseLong("98f2bc8e61", 16),
-        Long.parseLong("79b76d99e2", 16),
-        Long.parseLong("f33e5fb3c4", 16),
-        Long.parseLong("ae2eabe2a8", 16),
-        Long.parseLong("1e4f43e470", 16)};
+    /** The Bech32 character set for decoding. */
+    private static final byte[] CHARSET_REV = {
+            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
+            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
+            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
+            15, -1, 10, 17, 21, 20, 26, 30,  7,  5, -1, -1, -1, -1, -1, -1,
+            -1, 29, -1, 24, 13, 25,  9,  8, 23, -1, 18, 22, 31, 27, 19, -1,
+             1,  0,  3, 16, 11, 28, 12, 14,  6,  4,  2, -1, -1, -1, -1, -1,
+            -1, 29, -1, 24, 13, 25,  9,  8, 23, -1, 18, 22, 31, 27, 19, -1,
+             1,  0,  3, 16, 11, 28, 12, 14,  6,  4,  2, -1, -1, -1, -1, -1
+    };
 
-    private static final long POLYMOD_AND_CONSTANT = Long.parseLong("07ffffffff", 16);
-    private static final char[] CHARS = CHARSET.toCharArray();
+    public static class Bech32Data {
+        public final String hrp;
+        public final byte[] data;
 
-    private static Map<Character, Integer> charPositionMap;
-
-    static {
-        charPositionMap = new HashMap<>();
-        for (int i = 0; i < CHARS.length; i++) {
-            charPositionMap.put(CHARS[i], i);
-        }
-        if (charPositionMap.size() != 32) {
-            throw new RuntimeException("The charset must contain 32 unique characters.");
+        private Bech32Data(final String hrp, final byte[] data) {
+            this.hrp = hrp;
+            this.data = data;
         }
     }
 
-    /**
-     * Method calculateChecksumBytesPolymod calculates checksum from bitcoincash address
-     *
-     * @param checksumInput bitcoinhash address
-     * @return Returns a 40 bits checksum in form of 5 8-bit arrays. This still has
-     *         to me mapped to 5-bit array representation
-     */
-    public static byte[] calculateChecksumBytesPolymod(byte[] checksumInput) {
-        long l = 1;
-        for (int i = 0; i < checksumInput.length; i++) {
-            byte l0 = (byte)(l>>>35);
-            l = ((l & POLYMOD_AND_CONSTANT)<<5) ^ (Long.parseLong(String.format("%02x", checksumInput[i]), 16));
-            if ((l0 & 0x01) != 0) {
-                l = l ^ POLYMOD_GENERATORS[0].longValue();
+    /** Find the polynomial with value coefficients mod the generator as 30-bit. */
+    private static int polymod(final byte[] values) {
+        int c = 1;
+        for (byte v_i: values) {
+            int c0 = (c >>> 25) & 0xff;
+            c = ((c & 0x1ffffff) << 5) ^ (v_i & 0xff);
+            if ((c0 &  1) != 0) c ^= 0x3b6a57b2;
+            if ((c0 &  2) != 0) c ^= 0x26508e6d;
+            if ((c0 &  4) != 0) c ^= 0x1ea119fa;
+            if ((c0 &  8) != 0) c ^= 0x3d4233dd;
+            if ((c0 & 16) != 0) c ^= 0x2a1462b3;
+        }
+        return c;
+    }
+
+    /** Expand a HRP for use in checksum computation. */
+    private static byte[] expandHrp(final String hrp) {
+        int hrpLength = hrp.length();
+        byte ret[] = new byte[hrpLength * 2 + 1];
+        for (int i = 0; i < hrpLength; ++i) {
+            int c = hrp.charAt(i) & 0x7f; // Limit to standard 7-bit ASCII
+            ret[i] = (byte) ((c >>> 5) & 0x07);
+            ret[i + hrpLength + 1] = (byte) (c & 0x1f);
+        }
+        ret[hrpLength] = 0;
+        return ret;
+    }
+
+    /** Verify a checksum. */
+    private static boolean verifyChecksum(final String hrp, final byte[] values) {
+        byte[] hrpExpanded = expandHrp(hrp);
+        byte[] combined = new byte[hrpExpanded.length + values.length];
+        System.arraycopy(hrpExpanded, 0, combined, 0, hrpExpanded.length);
+        System.arraycopy(values, 0, combined, hrpExpanded.length, values.length);
+        return polymod(combined) == 1;
+    }
+
+    /** Create a checksum. */
+    private static byte[] createChecksum(final String hrp, final byte[] values)  {
+        byte[] hrpExpanded = expandHrp(hrp);
+        byte[] enc = new byte[hrpExpanded.length + values.length + 6];
+        System.arraycopy(hrpExpanded, 0, enc, 0, hrpExpanded.length);
+        System.arraycopy(values, 0, enc, hrpExpanded.length, values.length);
+        int mod = polymod(enc) ^ 1;
+        byte[] ret = new byte[6];
+        for (int i = 0; i < 6; ++i) {
+            ret[i] = (byte) ((mod >>> (5 * (5 - i))) & 31);
+        }
+        return ret;
+    }
+
+    /** Encode a Bech32 string. */
+    public static String encode(final Bech32Data bech32) {
+        return encode(bech32.hrp, bech32.data);
+    }
+
+    /** Encode a Bech32 string. */
+    public static String encode(String hrp, final byte[] values) {
+        if (hrp.length() < 1) {
+            throw new IllegalArgumentException("Human-readable part is too short");
+        }
+        if (hrp.length() > 83) {
+            throw new IllegalArgumentException("Human-readable part is too long");
+        }
+
+        hrp = hrp.toLowerCase(Locale.ROOT);
+        byte[] checksum = createChecksum(hrp, values);
+        byte[] combined = new byte[values.length + checksum.length];
+        System.arraycopy(values, 0, combined, 0, values.length);
+        System.arraycopy(checksum, 0, combined, values.length, checksum.length);
+        StringBuilder sb = new StringBuilder(hrp.length() + 1 + combined.length);
+        sb.append(hrp);
+        sb.append('1');
+        for (byte b : combined) {
+            sb.append(CHARSET.charAt(b));
+        }
+        return sb.toString();
+    }
+
+    /** Decode a Bech32 string. */
+    public static Bech32Data decode(final String str) throws AddressFormatException {
+        if (str.length() < 8) {
+            throw new AddressFormatException("Input too short: " + str.length());
+        }
+        if (str.length() > 90) {
+            throw new AddressFormatException("Input too long: " + str.length());
+        }
+        return decodeUnlimitedLength(str);
+    }
+
+    public static Bech32Data decodeUnlimitedLength(final String str) throws AddressFormatException {
+        boolean lower = false, upper = false;
+        for (int i = 0; i < str.length(); ++i) {
+            char c = str.charAt(i);
+            if (c < 33 || c > 126) throw new AddressFormatException("Invalid character: " + c + ", pos: "+ i);
+            if (c >= 'a' && c <= 'z') {
+                if (upper)
+                    throw new AddressFormatException("Invalid character: " + c + ", pos: "+ i);
+                lower = true;
             }
-
-            if ((l0 & 0x02) != 0) {
-                l = l ^ POLYMOD_GENERATORS[1].longValue();
-            }
-
-            if ((l0 & 0x04) != 0) {
-                l = l ^ POLYMOD_GENERATORS[2].longValue();
-            }
-
-            if ((l0 & 0x08) != 0) {
-                l = l ^ POLYMOD_GENERATORS[3].longValue();
-            }
-
-            if ((l0 & 0x10) != 0) {
-                l = l ^ POLYMOD_GENERATORS[4].longValue();
+            if (c >= 'A' && c <= 'Z') {
+                if (lower)
+                    throw new AddressFormatException("Invalid character: " + c + ", pos: "+ i);
+                upper = true;
             }
         }
-
-        byte[] checksum = long2Bytes((l ^ 1l));
-        if (checksum.length == 5) {
-            return checksum;
-        } else {
-            byte[] newChecksumArray = new byte[5];
-            System.arraycopy(checksum, Math.max(0, checksum.length - 5), newChecksumArray, Math.max(0, 5 - checksum.length), Math.min(5, checksum.length));
-            return newChecksumArray;
+        final int pos = str.lastIndexOf('1');
+        if (pos < 1) throw new AddressFormatException("Missing human-readable part");
+        final int dataPartLength = str.length() - 1 - pos;
+        if (dataPartLength < 6) throw new AddressFormatException("Data part too short: " + dataPartLength);
+        byte[] values = new byte[dataPartLength];
+        for (int i = 0; i < dataPartLength; ++i) {
+            char c = str.charAt(i + pos + 1);
+            if (CHARSET_REV[c] == -1) throw new AddressFormatException("Invalid character: " + c + ", pos: "+ (i + pos + 1));
+            values[i] = CHARSET_REV[c];
         }
+        String hrp = str.substring(0, pos).toLowerCase(Locale.ROOT);
+        if (!verifyChecksum(hrp, values)) throw new AddressFormatException("Invalid Checksum");
+        return new Bech32Data(hrp, Arrays.copyOfRange(values, 0, values.length - 6));
     }
-
-    /**
-     * Decode a base32 string back to the byte array representation
-     *
-     * @param base32String
-     * @return
-     */
-    public static byte[] decode(String base32String) {
-        byte[] bytes = new byte[base32String.length()];
-
-        char[] charArray = base32String.toCharArray();
-        for (int i = 0; i < charArray.length; i++) {
-            Integer position = charPositionMap.get(charArray[i]);
-            if (position == null) {
-                throw new RuntimeException("There seems to be an invalid char: " + charArray[i]);
-            }
-            bytes[i] = (byte) ((int) position);
-        }
-
-        return bytes;
-    }
-
-    /**
-     * Method concatenateByteArrays concatenates two byte arrays
-     *
-     * @param first - first byte array
-     * @param second - second byte array
-     * @return concatenated byte array
-     */
-    public static byte[] concatenateByteArrays(byte[] first, byte[] second) {
-        byte[] concatenatedBytes = new byte[first.length + second.length];
-        System.arraycopy(first, 0, concatenatedBytes, 0, first.length);
-        System.arraycopy(second, 0, concatenatedBytes, first.length, second.length);
-        return concatenatedBytes;
-    }
-
-    /**
-     * Method getPrefixBytes returns byte array for prefixString parameter
-     *
-     * @param prefixString prefix for bitcoincash
-     * @return byte array for bitcoincash
-     */
-    public static byte[] getPrefixBytes(String prefixString) {
-        byte[] prefixBytes = new byte[prefixString.length()];
-        char[] charArray = prefixString.toCharArray();
-        for (int i = 0; i < charArray.length; i++) {
-            prefixBytes[i] = (byte) (charArray[i] & 0x1f);
-        }
-
-        return prefixBytes;
-    }
-
-    /**
-     * Method calculateBitLength calculates the size of bits representing the number of long format
-     *
-     * @param value number of type long
-     * @return bit length of value
-     */
-    private static int calculateBitLength(long value) {
-        return Long.SIZE-Long.numberOfLeadingZeros(value);
-    }
-
-    /**
-     * Method long2Bytes converts long into byte array
-     *
-     * @param l number of type long
-     * @return byte array
-     */
-    private static byte[] long2Bytes(long l) {
-        int bitLength = calculateBitLength(l);
-        byte[] result = new byte[bitLength];
-        for (int i = bitLength - 1; i >= 0; i--) {
-            result[i] = (byte)(l & 0xFF);
-            l >>= bitLength;
-        }
-        return result;
-    }
-
-    /**
-     * Method bytes2Long converts byte array to long
-     *
-     * @param bytes byte array
-     * @return result of type long
-     */
-    public static long bytes2Long(byte[] bytes) {
-        long result = 0;
-        for (byte b : bytes) {
-            result <<= bytes.length;
-            result |= (b & 0xFF);
-        }
-        return result;
-    }
-
 }
